Martin front end

+++ frontend/src/components/Header.svelte
@@ -2,7 +2,8 @@
     /**
 	 * @type {string}
 	 */
-    export let headerStuff = "Test Header";
+    export let auth;
+    let isLoggedIn = auth.isAuth;
 </script>
 
 <style>
@@ -10,7 +11,7 @@
 
 <section>
     <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-        <a class="navbar-brand" href="/">{headerStuff}</a>
+        <a class="navbar-brand" href="/">Waterloo Housing</a>
         <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
         <span class="navbar-toggler-icon"></span>
         </button>
@@ -23,12 +24,19 @@
             <li class="nav-item active">
                 <a class="nav-link" href="/postings">Manage My Listings</a>
             </li>
-            <li class="nav-item me-auto">
-                <a class="nav-link me-auto" href="/login">Login</a>
-            </li>
-            <li class="nav-item me-auto">
-                <a class="nav-link me-auto" href="/delete">Delete Account</a>
-            </li>
+            {#if !isLoggedIn}
+                <li class="nav-item me-auto">
+                    <a class="nav-link me-auto" href="/login">Login</a>
+                </li>
+            {/if}
+            {#if isLoggedIn}
+                <li class="nav-item me-auto">
+                    <a class="nav-link me-auto" href="/logout">Log Out</a>
+                </li>
+                <li class="nav-item me-auto">
+                    <a class="nav-link me-auto" href="/delete">Delete Account</a>
+                </li>
+            {/if}
         </ul>
         </div>
     </nav>
